+++ test/with_dummyserver/test_poolmanager.py
@@ -109,34 +109,51 @@ class TestPoolManager(HTTPDummyServerTestCase):
         except MaxRetryError:
             pass
 
-    def test_redirect_cross_host_strip_auth_headers(self):
+    def test_redirect_cross_host_remove_headers(self):
         http = PoolManager()
         self.addCleanup(http.clear)
 
         r = http.request('GET', '%s/redirect' % self.base_url,
                          fields={'target': '%s/headers' % self.base_url_alt},
-                         headers={'Authentication': 'foo'})
+                         headers={'Authorization': 'foo'})
 
         self.assertEqual(r.status, 200)
 
         data = json.loads(r.data.decode('utf-8'))
 
-        self.assertNotIn('Authentication', data)
+        self.assertNotIn('Authorization', data)
 
-    def test_redirect_cross_host_forward_auth_headers(self):
+    def test_redirect_cross_host_no_remove_headers(self):
         http = PoolManager()
         self.addCleanup(http.clear)
 
         r = http.request('GET', '%s/redirect' % self.base_url,
                          fields={'target': '%s/headers' % self.base_url_alt},
-                         headers={'Authentication': 'foo'},
-                         retries=Retry(forward_auth_headers_across_hosts=True))
+                         headers={'Authorization': 'foo'},
+                         retries=Retry(remove_headers_on_redirect=[]))
 
         self.assertEqual(r.status, 200)
 
         data = json.loads(r.data.decode('utf-8'))
 
-        self.assertEqual(data['Authentication'], 'foo')
+        self.assertEqual(data['Authorization'], 'foo')
+
+    def test_redirect_cross_host_set_removed_headers(self):
+        http = PoolManager()
+        self.addCleanup(http.clear)
+
+        r = http.request('GET', '%s/redirect' % self.base_url,
+                         fields={'target': '%s/headers' % self.base_url_alt},
+                         headers={'X-API-Secret': 'foo',
+                                  'Authorization': 'bar'},
+                         retries=Retry(remove_headers_on_redirect=['X-API-Secret']))
+
+        self.assertEqual(r.status, 200)
+
+        data = json.loads(r.data.decode('utf-8'))
+
+        self.assertNotIn('X-API-Secret', data)
+        self.assertEqual(data['Authorization'], 'bar')
 
     def test_raise_on_redirect(self):
         http = PoolManager()